wenet_src/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.26.4
requests
tqdm
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
openai-whisper
librosa
pyyaml
jieba
langid
sentencepiece
deepspeed
tensorboardX
huggingface_hub
urllib3
awscli